requirements.txt
openmeteo-requests
requests-cache
retry-requests 
numpy
pandas
tensorflow-cpu
scikit-learn

fastapi
streamlit~=1.36.0